+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.4.2
 Time Zone API moved to Heroku
+Small memory optimization
 
 ## v1.4.1
 Mobile fixes